adding authorization

+++ src/backend/routes/posts.js
@@ -40,7 +40,8 @@ router.post("", checkAuth, upload , (req, res, next) => {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -54,23 +55,35 @@ router.post("", checkAuth, upload , (req, res, next) => {
 
 });
 
-router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
+router.put(
+  "/:id",
+  checkAuth,
+  multer({ storage: storage }).single("image"),
+   (req, res, next) => {
   let imagePath = req.body.imagePath;
   if(req.file) {
     const url = req.protocol + "://" + req.get("host");
     imagePath = url + "/images/" + req.file.filename
+    
   }
 
   const post = new Post({
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Update successful!" });
+  //console.log(post);
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId}, post).then(
+    result => {
+    if(result.nModified > 0) {
+
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
+
   });
 });
 
@@ -112,11 +125,16 @@ router.get("/:id", (req, res, next) => {
 
 router.delete("/:id", checkAuth, (req, res, next) => {
   //console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
     result => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted!" });
-    });
+      //console.log(result);
+      if (result.n > 0) {
+        res.status(200).json({ message: "Deletion successfull!" });
+      } else {
+        res.status(401).json({ message: "Not Authorized" });
+      }
+    }
+  );
 
 });
